Add files via upload
This program uses data structures, dynamic memory, arrays, menu driven development and pointers. The programs focus is to store insurance information into a mock 'database' for 3 users and then allow the user to grab the information from the struct. This was a programming assignment for my intermediate Computer Science class based in C++